==> scripts-cces-link/16_release-cces.R <==

# Save
pre  <- read_dta(path(output_dir, "RespondentLevel_Pre.dta"))
post <- read_dta(path(output_dir, "RespondentLevel_Post.dta"))

pre_fmt <- paste_labels(pre)
post_fmt <- paste_labels(post)

write_dta(pre_fmt, path(release_dir, "cces_candidates_pre.dta"))
write_dta(post_fmt, path(release_dir, "cces_candidates_post.dta"))
